remote_agent/requirements.txt
flask>=2.0
mss>=6.0
Pillow>=9.0
pywinauto>=0.6.8
